+++ pkg/reconciler/kubernetes/tektonconfig/controller.go
@@ -18,94 +18,24 @@ package tektonconfig
 
 import (
 	"context"
-	"os"
 
-	"github.com/go-logr/zapr"
-	mfc "github.com/manifestival/client-go-client"
-	mf "github.com/manifestival/manifestival"
-	"go.uber.org/zap"
 	"k8s.io/client-go/tools/cache"
 
 	"github.com/tektoncd/operator/pkg/apis/operator/v1alpha1"
-	operatorclient "github.com/tektoncd/operator/pkg/client/injection/client"
-	tektonConfiginformer "github.com/tektoncd/operator/pkg/client/injection/informers/operator/v1alpha1/tektonconfig"
 	tektonDashboardinformer "github.com/tektoncd/operator/pkg/client/injection/informers/operator/v1alpha1/tektondashboard"
-	tektonPipelineinformer "github.com/tektoncd/operator/pkg/client/injection/informers/operator/v1alpha1/tektonpipeline"
-	tektonTriggerinformer "github.com/tektoncd/operator/pkg/client/injection/informers/operator/v1alpha1/tektontrigger"
-	tektonConfigreconciler "github.com/tektoncd/operator/pkg/client/injection/reconciler/operator/v1alpha1/tektonconfig"
-	"github.com/tektoncd/operator/pkg/reconciler/common"
-	kubeclient "knative.dev/pkg/client/injection/kube/client"
-	deploymentinformer "knative.dev/pkg/client/injection/kube/informers/apps/v1/deployment"
+	"github.com/tektoncd/operator/pkg/reconciler/shared/tektonconfig"
 	"knative.dev/pkg/configmap"
 	"knative.dev/pkg/controller"
-	"knative.dev/pkg/injection"
-	"knative.dev/pkg/logging"
 )
 
 // NewController initializes the controller and is called by the generated code
 // Registers eventhandlers to enqueue events
 func NewController(ctx context.Context, cmw configmap.Watcher) *controller.Impl {
 
-	ctrl := NewExtendedController(KubernetesExtension)(ctx, cmw)
+	ctrl := tektonconfig.NewExtensibleController(KubernetesExtension)(ctx, cmw)
 	tektonDashboardinformer.Get(ctx).Informer().AddEventHandler(cache.FilteringResourceEventHandler{
 		FilterFunc: controller.FilterControllerGVK(v1alpha1.SchemeGroupVersion.WithKind("TektonConfig")),
 		Handler:    controller.HandleAll(ctrl.EnqueueControllerOf),
 	})
 	return ctrl
-}
-
-// NewExtendedController returns a controller extended to a specific platform
-func NewExtendedController(generator common.ExtensionGenerator) injection.ControllerConstructor {
-	return func(ctx context.Context, cmw configmap.Watcher) *controller.Impl {
-		tektonConfigInformer := tektonConfiginformer.Get(ctx)
-		tektonPipelineInformer := tektonPipelineinformer.Get(ctx)
-		tektonTriggerInformer := tektonTriggerinformer.Get(ctx)
-		deploymentInformer := deploymentinformer.Get(ctx)
-		kubeClient := kubeclient.Get(ctx)
-		logger := logging.FromContext(ctx)
-
-		mfclient, err := mfc.NewClient(injection.GetConfig(ctx))
-		if err != nil {
-			logger.Fatalw("Error creating client from injected config", zap.Error(err))
-		}
-		mflogger := zapr.NewLogger(logger.Named("manifestival").Desugar())
-		manifest, err := mf.ManifestFrom(mf.Slice{}, mf.UseClient(mfclient), mf.UseLogger(mflogger))
-		if err != nil {
-			logger.Fatalw("Error creating initial manifest", zap.Error(err))
-		}
-
-		c := &Reconciler{
-			kubeClientSet:     kubeClient,
-			operatorClientSet: operatorclient.Get(ctx),
-			extension:         generator(ctx),
-			manifest:          manifest,
-		}
-		impl := tektonConfigreconciler.NewImpl(ctx, c)
-
-		logger.Info("Setting up event handlers")
-
-		tektonConfigInformer.Informer().AddEventHandler(controller.HandleAll(impl.Enqueue))
-
-		tektonPipelineInformer.Informer().AddEventHandler(cache.FilteringResourceEventHandler{
-			FilterFunc: controller.FilterControllerGVK(v1alpha1.SchemeGroupVersion.WithKind("TektonConfig")),
-			Handler:    controller.HandleAll(impl.EnqueueControllerOf),
-		})
-
-		tektonTriggerInformer.Informer().AddEventHandler(cache.FilteringResourceEventHandler{
-			FilterFunc: controller.FilterControllerGVK(v1alpha1.SchemeGroupVersion.WithKind("TektonConfig")),
-			Handler:    controller.HandleAll(impl.EnqueueControllerOf),
-		})
-
-		deploymentInformer.Informer().AddEventHandler(cache.FilteringResourceEventHandler{
-			FilterFunc: controller.FilterControllerGVK(v1alpha1.SchemeGroupVersion.WithKind("TektonConfig")),
-			Handler:    controller.HandleAll(impl.EnqueueControllerOf),
-		})
-
-		if os.Getenv("AUTOINSTALL_COMPONENTS") == "true" {
-			// try to ensure that there is an instance of tektonConfig
-			newTektonConfig(operatorclient.Get(ctx), kubeclient.Get(ctx), manifest).ensureInstance(ctx)
-		}
-
-		return impl
-	}
 }
